Add some documentation and move command line utilities out of the crate lib
